# perf: Elide the total order wrapper for non-(float/option) types