--- a/crates/polars-arrow/src/array/primitive/mutable.rs
+++ b/crates/polars-arrow/src/array/primitive/mutable.rs
@@ -2,7 +2,6 @@ use std::iter::FromIterator;
 use std::sync::Arc;
 
 use polars_error::PolarsResult;
-use polars_utils::total_ord::TotalOrdWrap;
 
 use super::{check, PrimitiveArray};
 use crate::array::physical_binary::extend_validity;
@@ -364,14 +363,6 @@ impl<T: NativeType> Extend<Option<T>> for MutablePrimitiveArray<T> {
         let iter = iter.into_iter();
         self.reserve(iter.size_hint().0);
         iter.for_each(|x| self.push(x))
-    }
-}
-
-impl<T: NativeType> Extend<Option<TotalOrdWrap<T>>> for MutablePrimitiveArray<T> {
-    fn extend<I: IntoIterator<Item = Option<TotalOrdWrap<T>>>>(&mut self, iter: I) {
-        let iter = iter.into_iter();
-        self.reserve(iter.size_hint().0);
-        iter.for_each(|x| self.push(x.map(|x| x.0)))
     }
 }