--- a/crates/polars-ops/src/frame/join/mod.rs
+++ b/crates/polars-ops/src/frame/join/mod.rs
@@ -12,6 +12,7 @@ mod merge_sorted;
 #[cfg(feature = "chunked_ids")]
 use std::borrow::Cow;
 use std::fmt::{Debug, Display, Formatter};
+use std::hash::Hash;
 
 use ahash::RandomState;
 pub use args::*;