… crates/polars-ops/src/frame/join/hash_join/single_keys_inner.rs
… crates/polars-ops/src/frame/join/hash_join/single_keys_left.rs
… crates/polars-ops/src/frame/join/hash_join/single_keys_outer.rs
… crates/polars-ops/src/frame/join/hash_join/single_keys_semi_anti.rs
… crates/polars-ops/src/frame/pivot/positioning.rs
… crates/polars-ops/src/series/ops/approx_unique.rs
… crates/polars-ops/src/series/ops/is_first_distinct.rs
… crates/polars-ops/src/series/ops/is_in.rs
… crates/polars-ops/src/series/ops/is_last_distinct.rs
… crates/polars-ops/src/series/ops/is_unique.rs
… crates/polars-ops/src/series/ops/unique.rs
… crates/polars-plan/src/logical_plan/optimizer/simplify_expr.rs
… crates/polars-utils/src/total_ord.rs
… py-polars/tests/unit/dataframe/test_df.py