This should be a slight perf improvement for hashing primitive types, but more importantly will revert the change in the hash output values for integer types introduced by https://github.com/pola-rs/polars/pull/14617 (see https://github.com/pola-rs/polars/pull/14617#issuecomment-1960587029).

<details>
<summary>Benchmark</summary>

```
running 1 test
n_iterations: 70000000
(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): 2449ms vs 2157ms
(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): 2411ms vs 2151ms
(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): 2400ms vs 2122ms
(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): 2437ms vs 2145ms
(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): 2440ms vs 2166ms
(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2916ms vs 3357ms
(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 3051ms vs 3350ms
(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2921ms vs 3279ms
(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2961ms vs 3262ms
(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2989ms vs 3253ms
(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2399ms vs 2584ms
test total_ord::test::test has been running for over 60 seconds
(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2432ms vs 2665ms
(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2403ms vs 2665ms
(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2425ms vs 2578ms
(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): 2411ms vs 2681ms
```
```rust
mod test {
    #[test]
    fn test() {
        use std::time::SystemTime;

        use super::*;
        let rs = ahash::RandomState::new();

        let n_iterations = 70_000_000;

        println!("n_iterations: {}", n_iterations);

        fn time_it<F: Fn() -> T, T>(f: F, n: u32) -> u128 {
            let start = SystemTime::now();
            for _ in 0..n {
                f();
            }
            let end = SystemTime::now();
            let duration = end.duration_since(start).unwrap();
            return duration.as_millis();
        }

        let x: i32 = 10;

        for _ in 0..5 {
            let hash_one_wrap_t = time_it(|| rs.clone().hash_one(TotalOrdWrap(x)), n_iterations);
            let hash_one_to_tot_t = time_it(|| rs.clone().hash_one(x.to_total_ord()), n_iterations);
            println!(
                "(10 as i32) hash_one(TotalOrdWrap(x)) vs hash_one(x.to_total_ord()): {}ms vs {}ms",
                hash_one_wrap_t, hash_one_to_tot_t
            );
        }

        let x: Option<i32> = Some(10);

        for _ in 0..5 {
            let hash_one_t = time_it(|| rs.clone().hash_one(x), n_iterations);
            let hash_one_to_tot_t = time_it(|| rs.clone().hash_one(x.to_total_ord()), n_iterations);
            println!(
                "(Some(10) as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): {}ms vs {}ms",
                hash_one_t, hash_one_to_tot_t
            );
        }

        let x: Option<i32> = None;

        for _ in 0..5 {
            let hash_one_t = time_it(|| rs.clone().hash_one(x), n_iterations);
            let hash_one_to_tot_t = time_it(|| rs.clone().hash_one(x.to_total_ord()), n_iterations);
            println!(
                "(None as Option<i32>) hash_one(x) vs hash_one(x.to_total_ord()): {}ms vs {}ms",
                hash_one_t, hash_one_to_tot_t
            );
        }
    }
}
```
</details>